requirements.txt
flask==2.3.3
confluent-kafka==2.5.0
azure-storage-blob==12.23.1
azure-identity==1.19.0
gunicorn==21.2.0 